[Fix]: Added roles to TokenResponse and update token generation
Updated the `TokenResponse` record to include a new `string[] roles` property for storing user roles.
Modified `TokenService.cs` to retrieve user roles using `_userManager.GetRolesAsync(user)` and populate the `roles` property in the `TokenResponse` object.
Adjusted the assignment of the `Token` property to use either `jwtToken` or `newJwtToken` based on the context.